implemented loading multiple env files. Added logging.
