+++ src/RepoM.Api/IO/ExpressionEvaluator/RepositoryExpressionEvaluator.cs
@@ -8,26 +8,6 @@ using ExpressionStringEvaluator.Parser;
 using ExpressionStringEvaluator.VariableProviders;
 using RepoM.Api.Git;
 
-public class RepositoryContext
-{
-    public RepositoryContext()
-    {
-        Repositories = Array.Empty<Repository>();
-    }
-
-    public RepositoryContext(params Repository[] repositories)
-    {
-        Repositories = repositories.ToArray();
-    }
-
-    public RepositoryContext(IEnumerable<Repository> repositories)
-    {
-        Repositories = repositories.ToArray();
-    }
-
-    public Repository[] Repositories { get; }
-}
-
 public class RepositoryExpressionEvaluator
 {
     private readonly ExpressionExecutor _expressionExecutor;
